requirements.txt
Django==2.2.3
djangorestframework==3.10.1
psycopg2-binary==2.8.3
